Fix secrets masking in debug output
* Extract logic for escaping and unescaping  special symbols to `CommandStringConvertor` class
* Add escape data of `##vso` commands to the secret masker to resolve the problem with debug output masking
* Remove `Agent.Sdk.Knob` from `Command` class as it is not used in the code.
